Split `morton` into its own lib
